refactor: Move SelectMultiple to own file
